+++ src/components/molecules/MoreStories/style.tsx
@@ -1,0 +1,10 @@
+import styled from "styled-components";
+
+export const GridAndListToggleWrap = styled.div`
+  margin-left: 1rem;
+`;
+export const DittoWrap = styled.div`
+  width: 100%;
+  display: flex;
+  flex-wrap: wrap;
+`;
